backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
boto3==1.38.7
botocore==1.38.7
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
edge-tts==7.0.1
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-genai==1.13.0
googleapis-common-protos==1.70.0
groq==0.24.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
jmespath==1.0.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.7.0
motor==3.7.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.4.3
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
opencv-python==4.11.0.86
pillow==10.4.0
proglog==0.1.11
propcache==0.3.1
proto-plus==1.26.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pymongo==4.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.12.0
six==1.17.0
sniffio==1.3.1
srt==3.5.3
starlette==0.46.2
sympy==1.14.0
tabulate==0.9.0
tiktoken==0.9.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.20
uvicorn==0.34.2
websockets==15.0.1
yarl==1.20.0
